a-2025-job/quant-projects/equity_multi_factor/requirements.txt
pandas>=2.0
numpy>=1.25
pyarrow>=15.0
backtrader>=1.9.78.123
yfinance>=0.2.40
tqdm>=4.66
matplotlib>=3.8
python-dateutil>=2.9
requests>=2.32
